+++ example/src/bem-declarations.ts
@@ -1,15 +1,16 @@
+/* tslint:disable:no-console */
+import * as chalk from 'chalk';
+import { diffLines } from 'diff';
 import * as fs from 'fs';
 import * as path from 'path';
-import { diffLines } from 'diff';
-import * as chalk from 'chalk';
 import IDiffResult = JsDiff.IDiffResult;
-import { ENCODING, default as parseCss } from './parseCss';
+import { default as parseCss, ENCODING } from './parseCss';
 
 export default function getBemDeclarations(fileName: string) {
   console.log('Adding BEM declarations for file ' + path.join(__dirname, fileName));
   const { newCssFile, cssFile } = parseCss(fileName);
 
-  const diff: Array<IDiffResult> = diffLines(cssFile, newCssFile);
+  const diff: IDiffResult[] = diffLines(cssFile, newCssFile);
 
   diff.forEach((part: IDiffResult) => {
     const color = part.added ? 'green' : part.removed ? 'red' : 'grey';