+++ example/src/parseCss.ts
@@ -1,6 +1,6 @@
+import { parse, stringify, Stylesheet } from 'css';
 import * as fs from 'fs';
 import * as path from 'path';
-import { parse, stringify, Stylesheet } from 'css';
 import { preprocess } from '../../src';
 
 export const ENCODING: string = 'utf8';
@@ -9,7 +9,7 @@ export default function parseCss(fileName: string) {
   const cssFile: string = fs.readFileSync(path.join(__dirname, fileName), ENCODING);
   const css: Stylesheet = parse(cssFile);
 
-  preprocess(css, () => {});
+  preprocess(css, () => 'css');
 
   const newCssFile: string = stringify(css);